rbradford

This function generates a vector of values from the Standardized Bradford Distribution with length n and parameter c.

rbradford(n, c = 5)

Arguments

n
number of observations. If length(n)>1, the length is taken to be the number required.

c
parameter of the standardized bradford distribution. Must be greater than 0.

Value

a vector of length n with values generated from the Standardized Bradford Distribution with parameter c.

Examples

rbradford(n=100, c=10)

#>   [1] 0.021364760 0.639381061 0.322306720 0.045785967 0.001790143 0.205984041
#>   [7] 0.229899556 0.100337149 0.479719542 0.537526513 0.714334305 0.052118449
#>  [13] 0.008557210 0.115599305 0.162410787 0.059870795 0.163173189 0.016492127
#>  [19] 0.153974845 0.937357329 0.100397231 0.408698451 0.483118005 0.059980816
#>  [25] 0.949849014 0.491854694 0.013129476 0.256581971 0.430427336 0.421263343
#>  [31] 0.007776241 0.071750960 0.105723072 0.360055933 0.215393444 0.181877225
#>  [37] 0.444095371 0.872390501 0.054100296 0.068220122 0.410877395 0.230745672
#>  [43] 0.365843650 0.387096817 0.025892210 0.527033018 0.533189510 0.975772790
#>  [49] 0.924928027 0.154268218 0.202187302 0.112956282 0.052021915 0.257747656
#>  [55] 0.226640459 0.547987014 0.063166062 0.453256733 0.016927209 0.133812847
#>  [61] 0.623363583 0.092820106 0.292320175 0.123673944 0.317776211 0.058287083
#>  [67] 0.870497525 0.267227444 0.269101653 0.095042403 0.191872168 0.143718799
#>  [73] 0.006960271 0.205686072 0.154786160 0.004929057 0.146930518 0.282903323
#>  [79] 0.680837347 0.151615863 0.254624666 0.322181816 0.087150168 0.100473105
#>  [85] 0.216189013 0.808002504 0.161400894 0.066723378 0.400694762 0.015090720
#>  [91] 0.992296405 0.042956677 0.246753687 0.660577083 0.459758680 0.078362132
#>  [97] 0.271267525 0.640212652 0.006933357 0.208186334